+++ app/controllers/questions_controller.rb
@@ -1,21 +1,29 @@
 class QuestionsController < ApplicationController
   def index
-    questions = Question.all.order(created_at: "DESC")
+    questions = Question.joins(:user, :like).select('questions.*, users.id as user_id, users.name as user_name, likes.count as like_count, likes.id as like_id').order(created_at: "DESC")
     if questions
       render json: { "questions" => questions }
     else
-      render json: { message: "質問またはタグを受け取れませんでした。"}
+      render json: { message: "質問またはタグを受け取れませんでした。" }
     end
   end
 
   def show
-    question = Question.find(params[:id])
-    question_tags = question.tags
-    tags = Tag.all
+    question = Question.joins(:user, :like).select('questions.*, users.id as user_id, users.name as user_name, likes.count as like_count, likes.id as like_id').find(params[:id])
     if question
-      render json: { "question" => question}
+      render json: { "question" => question }
     else 
-      render json: { message: "質問または返信を受け取れませんでした。"}
+      render json: { message: "質問または返信を受け取れませんでした。" }
+    end
+  end
+
+  def user
+    puts current_user
+    questions = Question.joins(:user, :like).select('questions.*, users.id as user_id, users.name as user_name, likes.count as like_count, likes.id as like_id').where(user_id: current_user.id).order(created_at: "DESC")
+    if questions
+      render json: { get_my_questions: true, questions: questions}
+    else 
+      render json: { get_my_questions: false}
     end
   end
 
@@ -25,37 +33,38 @@ class QuestionsController < ApplicationController
     question = question_details[:question]
     target = Question.find_by(id: question_id)
     begin
-      target.update!(
-        title: question[:title],
-        content: question[:content],
-        anonymous: question[:anonymous]
-      )
-      puts "変更できました"
+      if target.user_id == current_user[:id] then
+        target.update!(
+          title: question[:title],
+          content: question[:content],
+          anonymous: question[:anonymous],
+          solved: question[:solved]
+        )
+        puts "変更できました"
+        render json: {update_question: true}
+      else
+        puts "投稿者以外は内容に変更を加えることができません"
+      end
     rescue ActiveRecord::RecordInvalid=> exception
       puts exception
       puts "変更できませんでした"   
+      render json: {update_question: false}
     end
   end
 
   def create
     details = receiveBody
     question = details[:question]
-    user = details[:user]
-    target = Question.new(
-      title: question[:title],
-      content: question[:content],
-      like: question[:like],
-      anonymous: question[:anonymous],
-      solved: question[:solved],
-      user_id: user[:id]
-    )
-
+    user = User.find(current_user[:id])
+    target = user.questions.new(question)
+    like = target.build_like(count: 0)
     begin
       target.save!
-      puts "保存成功"
+      like.save!
+      render json: {posted: true}
     rescue ActiveRecord::RecordInvalid => exception
       puts exception
-      puts "保存失敗"
+      render json: {posted: false}
     end
   end
 
@@ -63,9 +72,13 @@ class QuestionsController < ApplicationController
     question_id = params[:id]
     target = Question.find_by(id: question_id)
     begin
-      target.destroy!
-      puts "削除に成功しました"
-      # 保存成功時の処理
+      if target.user_id == current_user[:id] then
+        target.destroy!
+        puts "削除に成功しました"
+        # 保存成功時の処理
+      else
+        puts "投稿者以外は質問を削除することができません"
+      end
     rescue ActiveRecord::RecordInvalid => e
       puts e
       puts "削除に失敗しました"
